+++ example/lib/cupertino_form.dart
@@ -1,5 +1,3 @@
-import 'package:flrx_validator/rules/email_rule.dart';
-import 'package:flrx_validator/rules/required_rule.dart';
 import 'package:flrx_validator/validator.dart';
 import 'package:flutter/cupertino.dart';