+++ test/unit/utils/string_utils_test.dart
@@ -1,4 +1,4 @@
-import 'package:flrx_validator/src/string_utils.dart';
+import 'package:flrx_validator/src/utils/string_utils.dart';
 import 'package:test/test.dart';
 
 void main() {